Restructure and package PlanTUS
