+++ PlanTUS_wrapper.py
@@ -3,46 +3,14 @@
 """
 PlanTUS wrapper script
 ======================
-This script orchestrates the PlanTUS pipeline using the functions in
-`PlanTUS.py`. Its inputs are the subject-specific data paths and device
-parameters near the top of the file.
+This script is just a wrapper script around PlanTUS_cli to keep backwards
+compatibility and for BabelBrain usage.
 
-Usage
------
-1. Fill in the *Subject-specific variables* and *Transducer-specific
-   variables* below.
-2. Run this script with Python 3.
+(The parameters from PlanTUS_cli will change in the future).
 
-Outputs
--------
-All outputs are written into `<m2m_subject>/PlanTUS/<ROI-name>/` and
-include surfaces (`*.surf.gii`/`*.stl`) and per-vertex metrics
-(`*.func.gii`).
-
-
-(c) Maximilian Lueckel, mlueckel@uni-mainz.de
 """
 import argparse
-import platform
-import os
-import sys
-from pathlib import Path
-import yaml
-
-_IS_MAC = platform.system() == 'Darwin'
-
-#handy code to recover current path of Python script, works with Pyinstaller
-def resource_path():  # needed for bundling
-    """Get absolute path to resource, works for dev and for PyInstaller"""
-    if not _IS_MAC:
-        return os.path.split(Path(__file__))[0]
-
-    if getattr(sys, 'frozen', False) and hasattr(sys, '_MEIPASS'):
-        bundle_dir = Path(sys._MEIPASS)
-    else:
-        bundle_dir = Path(__file__).parent
-
-    return bundle_dir
+from PlanTUS.PlanTUS_cli import run
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(description="PlanTUS wrapper")
@@ -50,593 +18,11 @@ if __name__ == "__main__":
     parser.add_argument("mesh", type=str,  help="Path to head mesh")
     parser.add_argument("roi", type=str, help="Path to target ROI")
     parser.add_argument("config", type=str, help="Path to configuration file")
-    parser.add_argument( "--skip_wb_view",action="store_true",help="Run calculations but skip wb_view")
-    parser.add_argument( "--use_internal_viewer",action="store_true",help="Use own viewer instead of wb_view")
-    parser.add_argument( "--do_only_trajectory",type=int,default=-1,help="Optional integer to run only the generation of trajectory (default: -1). Specify number of triangles to generate."
-)
+    parser.add_argument("--skip_wb_view",action="store_true",help="Run calculations but skip wb_view")
+    parser.add_argument("--use_internal_viewer",action="store_true",help="Use own viewer instead of wb_view")
+    parser.add_argument("--do_only_trajectory",type=int,default=-1,help="Optional integer to run only the generation of trajectory (default: -1). Specify number of triangles to generate.")
 
     args = parser.parse_args()
-#===============================================================================
-#===============================================================================
-# Configure inputs
-#===============================================================================
-#===============================================================================
 
-# Subject-specific variables ---------------------------------------------------
+    run(args)
 
-# Path to T1 image (output for SimNIBS' charm)
-t1_filepath = args.t1
-
-# Path to head mesh (.msh file generated by SimNIBS' charm)
-simnibs_mesh_filepath = args.mesh
-
-# Path to mask of target region of interest (in same space as T1 image)
-# Note: PlanTUS output folder will have same name as this file
-target_roi_filepath = args.roi
-
-
-# Transducer-specific variables ------------------------------------------------
-with open(args.config, 'r') as file:
-    config = yaml.safe_load(file)
-
-IDTarget = config.get("IDTarget", "")
-
-# Maximum minimum, and optimal focal depth of transducer (in mm)
-max_distance = config["max_distance"]
-min_distance = config["min_distance"]
-optimal_distance = config["optimal_distance"]
-
-# Aperture diameter (in mm)
-transducer_diameter = config["transducer_diameter"]
-
-# Maximum allowed angle for tilting of TUS transducer (in degrees)
-max_angle = config["max_angle"]
-
-# Offset between radiating surface and exit plane of transducer (in mm)
-plane_offset = config["plane_offset"]
-
-# Additional offset between skin and exit plane of transducer (in mm;
-# e.g., due to addtional gel/silicone pad)
-additional_offset = config["additional_offset"]
-
-# Sampled of focal distance and corresponding FLHM values (both in mm)
-#according to, e.g., calibration report
-focal_distance_list = config["focal_distance_list"]
-flhm_list = config["flhm_list"]
-
-
-# User-defined metric weights --------------------------------------------------
-# Note: each in [0,1], sum = 1
-weight_skin_target_distances = config["weight_skin_target_distances"]
-weight_skin_target_angles = config["weight_skin_target_angles"]
-weight_skin_target_intersections = config["weight_skin_target_intersections"]
-weight_skin_skull_angles = config["weight_skin_skull_angles"]
-weight_skull_thickness = config["weight_skull_thickness"]
-
-# PlanTUS paths ----------------------------------------------------------------
-plantus_main_folder = str(resource_path())
-plantus_code_path = os.path.join(plantus_main_folder, "code")
-planning_scene_template_filepath = os.path.join(plantus_main_folder, "resources", "scene_templates", "TUSTransducerPlacementPlanning_TEMPLATE.scene")
-placement_scene_template_filepath = os.path.join(plantus_main_folder, "resources", "scene_templates", "TUSTransducerPlacement_TEMPLATE.scene")
-# Transducer model (*.surf.gii) — leave empty to generate a generic one
-bUseGenericTransducerModel = config.get("bUseGenericTransducerModel", False)
-if bUseGenericTransducerModel:
-        transducer_surface_model_filepath = ""
-else:
-    transducer_surface_model_filepath = os.path.join(plantus_main_folder, "resources", "transducer_models", "TRANSDUCER-NEUROFUS-CTX-500-4_DEVICE.surf.gii")
-# update paths according to user configuration
-connectome_wb_path = config["connectome_wb_path"]
-
-#===============================================================================
-# Setup
-#===============================================================================
-import sys
-import shutil
-import numpy as np
-import math
-sys.path.append(plantus_code_path)
-os.chdir(plantus_code_path)
-import PlanTUS as PlanTUS
-
-PlanTUS.set_paths(connectome_wb_path=connectome_wb_path)
-
-
-# Prepare output directory named by ROI
-roi_fname = os.path.split(target_roi_filepath)[1]
-roi_name = roi_fname.replace(".nii", "").replace(".gz", "")
-output_path = os.path.join(os.path.split(simnibs_mesh_filepath)[0], "PlanTUS", roi_name)
-os.makedirs(output_path, exist_ok=True)
-shutil.copy(target_roi_filepath, output_path)
-target_roi_filepath = os.path.join(output_path, roi_fname)
-
-# -----------------------------------------------------------------------------
-# Transducer model creation (if no pre-made model provided)
-# -----------------------------------------------------------------------------
-if transducer_surface_model_filepath == "":
-    # Generate a simple generic transducer surface model
-    transducer_surface_model_filepath = os.path.join(output_path, "generic_transducer.surf.gii")
-    PlanTUS.create_surface_transducer_model(transducer_diameter/2, plane_offset + additional_offset, transducer_surface_model_filepath)
-
-if args.do_only_trajectory<0:
-    # -----------------------------------------------------------------------------
-    # Convert SimNIBS meshes to surfaces (STL + GIFTI); annotate structures
-    # -----------------------------------------------------------------------------
-
-    # Skin
-    PlanTUS.convert_simnibs_mesh_to_surfaces(simnibs_mesh_filepath, [1005], "skin", output_path)
-    PlanTUS.add_structure_information(output_path + "/skin.surf.gii", "CORTEX_LEFT")
-
-    # Skull
-    PlanTUS.convert_simnibs_mesh_to_surfaces(simnibs_mesh_filepath, [1007], "skull", output_path)
-    PlanTUS.add_structure_information(output_path + "/skull.surf.gii", "CORTEX_RIGHT")
-
-
-    # -----------------------------------------------------------------------------
-    # Avoidance mask on the skin surface
-    # -----------------------------------------------------------------------------
-    avoidance_mask = PlanTUS.create_avoidance_mask(simnibs_mesh_filepath,
-                                                   os.path.join(output_path, "skin.surf.gii"),
-                                                   transducer_diameter / 2.5,
-                                                   )
-
-
-    # -----------------------------------------------------------------------------
-    # Metrics: distances, angles, intersections
-    # -----------------------------------------------------------------------------
-    # Distances skin → target center
-    center = PlanTUS.roi_center_of_gravity(target_roi_filepath)
-    skin_target_distances = PlanTUS.distance_between_surface_and_point(os.path.join(output_path, "skin.surf.gii"), center)
-    PlanTUS.create_metric_from_pseudo_nifti("distances", skin_target_distances, os.path.join(output_path, "skin.surf.gii"))
-    PlanTUS.mask_metric(os.path.join(output_path, "distances_skin.func.gii"), os.path.join(output_path, "avoidance_skin.func.gii"))
-    PlanTUS.add_structure_information(os.path.join(output_path, "distances_skin.func.gii"), "CORTEX_LEFT")
-    PlanTUS.threshold_metric(os.path.join(output_path, "distances_skin.func.gii"), max_distance)
-    PlanTUS.mask_metric(os.path.join(output_path, "distances_skin_thresholded.func.gii"), os.path.join(output_path, "avoidance_skin.func.gii"))
-    PlanTUS.add_structure_information(os.path.join(output_path, "distances_skin_thresholded.func.gii"), "CORTEX_LEFT")
-
-
-    # Angles between surface normals and skin→target vectors
-    angles = []
-    _, skin_normals = PlanTUS.compute_surface_metrics(os.path.join(output_path, "skin.surf.gii"))
-    skin_vecs = PlanTUS.vectors_between_surface_and_point(os.path.join(output_path, "skin.surf.gii"), center)
-    for i in range(len(skin_vecs)):
-        angles.append(math.degrees(PlanTUS.angle_between_vectors(skin_vecs[i], skin_normals[i])))
-    angles = np.abs(np.asarray(angles))
-    PlanTUS.create_metric_from_pseudo_nifti("angles", angles, os.path.join(output_path, "skin.surf.gii"))
-    PlanTUS.mask_metric(os.path.join(output_path, "angles_skin.func.gii"), os.path.join(output_path, "avoidance_skin.func.gii"))
-    PlanTUS.add_structure_information(os.path.join(output_path, "angles_skin.func.gii"), "CORTEX_LEFT")
-
-
-    # Intersections between normals and the target mesh
-    PlanTUS.stl_from_nii(target_roi_filepath, 0.25)
-    skin_coords, skin_normals = PlanTUS.compute_surface_metrics(os.path.join(output_path, "skin.surf.gii"))
-    skin_target_hits = PlanTUS.compute_vector_mesh_intersections(skin_coords, skin_normals, os.path.join(output_path, f"{roi_name}_3Dmodel.stl"), 200)
-
-    hit_vals = []
-    for hits in skin_target_hits:
-        if len(hits) == 1:
-            hit_vals.append(0)
-        elif len(hits) in (2, 3):
-            d = np.linalg.norm(np.asarray(hits[1]) - np.asarray(hits[0]))
-            hit_vals.append(d)
-        elif len(hits) == 4:
-            d = np.linalg.norm(np.asarray(hits[1]) - np.asarray(hits[0])) + np.linalg.norm(np.asarray(hits[3]) - np.asarray(hits[2]))
-            hit_vals.append(d)
-        elif len(hits) > 4:
-            hit_vals.append(np.nan)
-        else:
-            hit_vals.append(0)
-
-    target_intersection = np.asarray(hit_vals)
-    PlanTUS.create_metric_from_pseudo_nifti("target_intersection", target_intersection, os.path.join(output_path, "skin.surf.gii"))
-    PlanTUS.mask_metric(os.path.join(output_path, "target_intersection_skin.func.gii"), os.path.join(output_path, "avoidance_skin.func.gii"))
-    PlanTUS.add_structure_information(os.path.join(output_path, "target_intersection_skin.func.gii"), "CORTEX_LEFT")
-
-
-    # Skin–skull angle
-    skin_coords, skin_normals = PlanTUS.compute_surface_metrics(os.path.join(output_path, "skin.surf.gii"))
-    skull_coords, skull_normals = PlanTUS.compute_surface_metrics(os.path.join(output_path, "skull.surf.gii"))
-    skin_skull_hits = PlanTUS.compute_vector_mesh_intersections(skin_coords, skin_normals, os.path.join(output_path, "skull.stl"), 40)
-
-    idx_closest = []
-    for i in range(len(skin_coords)):
-        try:
-            p = skin_skull_hits[i][0]
-            dists = np.linalg.norm((skull_coords - p), axis=1)
-            idx_closest.append(int(np.argmin(dists)))
-        except Exception:
-            idx_closest.append(np.nan)
-    idx_closest = np.asarray(idx_closest).astype(int)
-
-    skin_skull_angles = []
-    for i in range(len(skin_coords)):
-        try:
-            ang = math.degrees(PlanTUS.angle_between_vectors(skin_normals[i], skull_normals[idx_closest[i]]))
-            skin_skull_angles.append(ang)
-        except Exception:
-            skin_skull_angles.append(0)
-    skin_skull_angles = np.asarray(skin_skull_angles)
-
-    PlanTUS.create_metric_from_pseudo_nifti("skin_skull_angles", skin_skull_angles, os.path.join(output_path, "skin.surf.gii"))
-    PlanTUS.mask_metric(os.path.join(output_path, "skin_skull_angles_skin.func.gii"), os.path.join(output_path, "avoidance_skin.func.gii"))
-    PlanTUS.add_structure_information(os.path.join(output_path, "skin_skull_angles_skin.func.gii"), "CORTEX_LEFT")
-
-
-    # Skull→target intersections projected on skin (aux metric)
-    skull_coords, skull_normals = PlanTUS.compute_surface_metrics(os.path.join(output_path, "skull.surf.gii"))
-    skull_target_hits = PlanTUS.compute_vector_mesh_intersections(skull_coords, skull_normals, os.path.join(output_path, f"{roi_name}_3Dmodel.stl"), 200)
-    non_empty_idx = [i for i, h in enumerate(skull_target_hits) if h]
-    skull_coords_in = skull_coords[non_empty_idx]
-    skull_normals_in = skull_normals[non_empty_idx]
-    skull_to_skin_hits = PlanTUS.compute_vector_mesh_intersections(skull_coords_in, skull_normals_in, os.path.join(output_path, "skin.stl"), -500)
-
-    all_xy = np.vstack([np.array(h) for h in skull_to_skin_hits if h])
-    intersection_coords = np.round(all_xy, 2)
-    mesh_round = np.round(skin_coords, 2)
-
-    matching = [np.where(np.all(np.isclose(mesh_round, c, atol=2), axis=1))[0] for c in intersection_coords]
-    matching = [m[0] for m in matching if len(m) > 0]
-    mask_arr = np.zeros(len(skin_coords), dtype=int)
-    mask_arr[matching] = 1
-
-    skull_target_intersection2 = np.where(mask_arr == 1, 1, 0)
-
-    PlanTUS.create_metric_from_pseudo_nifti("skin_skull_target_intersection", skull_target_intersection2, os.path.join(output_path, "skin.surf.gii"))
-    PlanTUS.mask_metric(os.path.join(output_path, "skin_skull_target_intersection_skin.func.gii"), os.path.join(output_path, "avoidance_skin.func.gii"))
-    PlanTUS.add_structure_information(os.path.join(output_path, "skin_skull_target_intersection_skin.func.gii"), "CORTEX_LEFT")
-
-
-    # Skull thickness (outer→inner or outer→outer where needed)
-    PlanTUS.convert_simnibs_mesh_to_surfaces(simnibs_mesh_filepath, [1007], "skull", output_path)
-    PlanTUS.add_structure_information(os.path.join(output_path, "skull.surf.gii"), "CORTEX_RIGHT")
-    PlanTUS.convert_simnibs_mesh_to_surfaces(simnibs_mesh_filepath, [1001, 1002, 1003, 1009], "skull_inner_surface", output_path)
-
-    skull_coords, skull_normals = PlanTUS.compute_surface_metrics(os.path.join(output_path, "skull.surf.gii"))
-    skull_inner_hits = PlanTUS.compute_vector_mesh_intersections(skull_coords, skull_normals, os.path.join(output_path, "skull_inner_surface.stl"), 100)
-
-    inner_nonempty = [i for i, h in enumerate(skull_inner_hits) if h]
-    inner_empty = [i for i, h in enumerate(skull_inner_hits) if not h]
-
-    inner_closest = np.asarray([skull_inner_hits[i][0] for i in inner_nonempty])
-
-    skull_outer_hits = PlanTUS.compute_vector_mesh_intersections(skull_coords, skull_normals, os.path.join(output_path, "skull.stl"), 100)
-    outer_hits_noinner = [skull_outer_hits[i] for i in inner_empty]
-    outer_closest = []
-    for h in outer_hits_noinner:
-        if len(h) == 0:
-            outer_closest.append(np.array([np.nan, np.nan, np.nan]))
-        elif len(h) == 1:
-            outer_closest.append(np.asarray(h[0]))
-        else:
-            outer_closest.append(np.asarray(h[1]))
-    outer_closest = np.asarray(outer_closest)
-
-    skull_coords_in = skull_coords[inner_nonempty]
-    skull_coords_out = skull_coords[inner_empty]
-
-    d_in = np.linalg.norm(skull_coords_in - inner_closest, axis=1)
-    d_out = np.linalg.norm(skull_coords_out - outer_closest, axis=1)
-
-    skull_thickness = np.zeros(len(skull_coords))
-    skull_thickness[inner_nonempty] = d_in
-    skull_thickness[inner_empty] = d_out
-
-    PlanTUS.create_metric_from_pseudo_nifti("skull_thickness", skull_thickness, os.path.join(output_path, "skull.surf.gii"))
-    PlanTUS.add_structure_information(os.path.join(output_path, "skull_thickness_skull.func.gii"), "CORTEX_RIGHT")
-
-    # Project thickness to skin via nearest skull vertex indices
-    skin_closest_thick = np.zeros(len(idx_closest))
-    for i in range(len(idx_closest)):
-        try:
-            skin_closest_thick[i] = skull_thickness[idx_closest[i]]
-        except Exception:
-            skin_closest_thick[i] = np.nan
-
-    PlanTUS.create_metric_from_pseudo_nifti("skull_thickness", skin_closest_thick, os.path.join(output_path, "skin.surf.gii"))
-    PlanTUS.mask_metric(os.path.join(output_path, "skull_thickness_skin.func.gii"), os.path.join(output_path, "avoidance_skin.func.gii"))
-    PlanTUS.add_structure_information(os.path.join(output_path, "skull_thickness_skin.func.gii"), "CORTEX_LEFT")
-
-
-    # -----------------------------------------------------------------------------
-    # Composite utility metric (geometric mean of normalized components)
-    # -----------------------------------------------------------------------------
-    # Zero-out invalid vertices using distance/avoidance
-    skin_target_distances_m = skin_target_distances.copy()
-    skin_target_distances_m[skin_target_distances > max_distance] = 0
-    skin_target_distances_m[avoidance_mask == 0] = 0
-
-    angles_m = angles.copy()
-    angles_m[skin_target_distances > max_distance] = 0
-    angles_m[avoidance_mask == 0] = 0
-
-    intersections_m = target_intersection.copy()
-    intersections_m[skin_target_distances > max_distance] = 0
-    intersections_m[avoidance_mask == 0] = 0
-
-    skin_skull_angles_m = skin_skull_angles.copy()
-    skin_skull_angles_m[skin_target_distances > max_distance] = 0
-    skin_skull_angles_m[avoidance_mask == 0] = 0
-
-    skin_closest_thick_m = skin_closest_thick.copy()
-    skin_closest_thick_m[skin_target_distances > max_distance] = 0
-    skin_closest_thick_m[avoidance_mask == 0] = 0
-
-    # Exponents (shape)
-    p_d = p_L = p_theta = p_delta = p_t = 1.0
-
-    # Weights
-    w_d = float(weight_skin_target_distances)
-    w_L = float(weight_skin_target_intersections)
-    w_theta = float(weight_skin_target_angles)
-    w_delta = float(weight_skin_skull_angles)
-    w_t = float(weight_skull_thickness)
-
-    # Distance utility (best at optimal, worst at max)
-    d_best, d_worst = float(optimal_distance), float(max_distance)
-    with np.errstate(divide='ignore', invalid='ignore'):
-        u_d = np.clip(((d_best - skin_target_distances_m) / (d_best - d_worst)) ** p_d, 0, 1)
-        u_d[~np.isfinite(u_d)] = 0
-
-    # Intersection utility (more is better)
-    L = intersections_m.astype(float)
-    L[np.isnan(L)] = 0
-    L_min = np.nanmin(L)
-    L_max = np.nanmax(L)
-    if L_max > L_min:
-        u_L = np.clip(((L - L_min) / (L_max - L_min)) ** p_L, 0, 1)
-    else:
-        u_L = np.zeros_like(L)
-
-    # Tilt utility (smaller angles are better)
-    A = angles_m.astype(float)
-    A[np.isnan(A)] = 0
-    theta_max = np.nanmax(A)
-    if theta_max > 0:
-        u_theta = (1 - A / theta_max) ** p_theta
-    else:
-        u_theta = np.ones_like(A)
-
-    # Alignment utility (blend intersections & tilt)
-    kappa = 0.7
-    w_align = w_L + w_theta
-    if w_align > 0:
-        rho = w_theta / w_align
-        u_align = 1 - np.power(1 - u_L, 1 - rho) * np.power(1 - kappa * u_theta, rho)
-    else:
-        u_align = np.ones_like(u_L)
-
-    # Skin–skull angle utility (smaller better)
-    D = skin_skull_angles_m.astype(float)
-    D[np.isnan(D)] = 0
-    delta_worst = np.nanmax(D)
-    if delta_worst > 0:
-        u_delta = np.clip((1 - D / delta_worst) ** p_delta, 0, 1)
-    else:
-        u_delta = np.ones_like(D)
-
-    # Skull thickness utility (thinner better)
-    T = skin_closest_thick_m.astype(float)
-    T[np.isnan(T)] = 0
-    t_best = np.nanmin(T)
-    t_worst = np.nanmax(T)
-    if t_worst > t_best:
-        u_t = np.clip((1 - (T - t_best) / (t_worst - t_best)) ** p_t, 0, 1)
-    else:
-        u_t = np.ones_like(T)
-
-    # Geometric aggregation
-    composite_metric = (np.power(u_align, w_align) * np.power(u_d, w_d) * np.power(u_delta, w_delta) * np.power(u_t, w_t))
-
-    # Persist composite metric
-    metric_base = (
-    "composite_"
-    f"TargetDistance{weight_skin_target_distances}_"
-    f"TargetAngle{weight_skin_target_angles}_"
-    f"TargetIntersection{weight_skin_target_intersections}_"
-    f"SkinSkullAngle{weight_skin_skull_angles}_"
-    f"SkullThickness{weight_skull_thickness}"
-    )
-
-    PlanTUS.create_metric_from_pseudo_nifti(metric_base, composite_metric, os.path.join(output_path, "skin.surf.gii"))
-    PlanTUS.mask_metric(os.path.join(output_path, metric_base + "_skin.func.gii"), os.path.join(output_path, "avoidance_skin.func.gii"))
-    PlanTUS.add_structure_information(os.path.join(output_path, metric_base + "_skin.func.gii"), "CORTEX_LEFT")
-
-
-
-    # ---------------------------------------------------------------
-    if args.skip_wb_view:
-            sys.exit(0)  #skip wb_view part if requested
-
-    if args.use_internal_viewer:
-        from PyQt5.QtWidgets import QApplication,QDialog,QVBoxLayout
-        from Viewer import PrepareShowResults
-
-        app = QApplication(sys.argv)
-
-        DlgResults=QDialog()
-        DlgResults.setWindowTitle("PlanTUS Results")
-
-        layout = QVBoxLayout()
-        DlgResults.setLayout(layout)
-        DlgResults.resize(1700, 700)
-
-        def CallBackGenerateTrajectory(selection):
-            DlgResults.accept()
-
-        WidgetViewer = PrepareShowResults(output_path +os.sep + "skin.surf.gii",
-                                          output_path +os.sep + "distances_skin.func.gii",
-                                          output_path +os.sep + "distances_skin_thresholded.func.gii",
-                                          output_path +os.sep + "target_intersection_skin.func.gii",
-                                          output_path +os.sep + "angles_skin.func.gii",
-                                          output_path +os.sep + "skin_skull_angles_skin.func.gii",
-                                          CallBackGenerateTrajectory=CallBackGenerateTrajectory)
-
-        WidgetViewer.resize(1700, 700)
-        layout.addWidget(WidgetViewer)
-
-        if not DlgResults.exec():
-            sys.exit(0) #we stop here
-        print('Generating trajectory for ID', WidgetViewer.select_vortex)
-        PlanTUS.prepare_acoustic_simulation(WidgetViewer.select_vortex,
-                                                                output_path,
-                                                                target_roi_filepath,
-                                                                t1_filepath,
-                                                                max_distance,
-                                                                min_distance,
-                                                                transducer_diameter,
-                                                                max_angle,
-                                                                plane_offset,
-                                                                additional_offset,
-                                                                transducer_surface_model_filepath,
-                                                                focal_distance_list,
-                                                                flhm_list,
-                                                                placement_scene_template_filepath,
-                                                                ID=IDTarget,
-                                                                use_internal_viewer=args.use_internal_viewer)
-
-
-
-    else:
-        # -----------------------------------------------------------------------------
-        # Create and open scences in Connectome Workbench viewer
-        # -----------------------------------------------------------------------------
-    
-        import subprocess
-        import re
-        import threading
-        from pynput import mouse
-    
-        scene_variable_names = [
-            'SKIN_SURFACE_FILENAME',
-            'SKIN_SURFACE_FILEPATH',
-            'SKULL_SURFACE_FILENAME',
-            'SKULL_SURFACE_FILEPATH',
-            'DISTANCES_FILENAME',
-            'DISTANCES_FILEPATH',
-            'INTERSECTION_FILENAME',
-            'INTERSECTION_FILEPATH',
-            'ANGLES_FILENAME',
-            'ANGLES_FILEPATH',
-            'ANGLES_SKIN_SKULL_FILENAME',
-            'ANGLES_SKIN_SKULL_FILEPATH',
-            'DISTANCES_MAX_FILENAME',
-            'DISTANCES_MAX_FILEPATH',
-            'T1_FILENAME',
-            'T1_FILEPATH',
-            'MASK_FILENAME',
-            'MASK_FILEPATH']
-    
-        scene_variable_values = [
-            'skin.surf.gii',
-            './skin.surf.gii',
-            'skull.surf.gii',
-            './skull.surf.gii',
-            'distances_skin.func.gii',
-            './distances_skin.func.gii',
-            'target_intersection_skin.func.gii',
-            './target_intersection_skin.func.gii',
-            'angles_skin.func.gii',
-            './angles_skin.func.gii',
-            'skin_skull_angles_skin.func.gii',
-            './skin_skull_angles_skin.func.gii',
-            'distances_skin_thresholded.func.gii',
-            './distances_skin_thresholded.func.gii',
-            'T1.nii.gz',
-            '../../T1.nii.gz',
-            roi_fname,
-            './' + roi_fname]
-
-
-    PlanTUS.create_scene(planning_scene_template_filepath, output_path + "/scene.scene", scene_variable_names, scene_variable_values)
-
-    # Define the command
-    command = connectome_wb_path + os.sep + "wb_view -logging FINER " + output_path + os.sep + "scene.scene"
-
-    # Regular expression pattern to match the phrase and the number
-    pattern = re.compile(r"Switched vertex to triangle nearest vertex\s+(\.\d+)")
-
-    # Initialize the variable to store the number and a flag to trigger processing
-    triangle_number = None
-    process_line = False
-
-    # Function to monitor mouse clicks
-    def on_click(x, y, button, pressed):
-        global process_line
-        if pressed:
-            process_line = True
-
-    # Start listening for mouse clicks in a separate thread
-    listener = mouse.Listener(on_click=on_click)
-    listener.start()
-
-    # Start the process to run the command
-    process = subprocess.Popen(command, stdout=subprocess.PIPE, stderr=subprocess.PIPE, shell=True, cwd=output_path, text=True)
-
-    # Function to read the process output
-    def read_output():
-        global triangle_number, process_line, IDTarget
-
-        while True:
-            output = process.stderr.readline()
-            if output == '' and process.poll() is not None:
-                break
-
-            if process_line:
-                # Process the latest line only when a mouse click is detected
-                match = pattern.search(output)
-                if match:
-                    triangle_number = match.group(1)
-                    triangle_number = int(triangle_number.replace(".", ""))
-                    print(f"Switched vertex to triangle nearest vertex: {triangle_number}")
-
-                    # Ask the user if they want to generate the transducer placement
-                    response = input(f"Generate transducer placement for vertex {triangle_number}? (yes/no): ").strip().lower()
-                    if response == "yes":
-                        print(f"Generating transducer placement for vertex {triangle_number}")
-                        PlanTUS.prepare_acoustic_simulation(triangle_number,
-                                                            output_path,
-                                                            target_roi_filepath,
-                                                            t1_filepath,
-                                                            max_distance,
-                                                            min_distance,
-                                                            transducer_diameter,
-                                                            max_angle,
-                                                            plane_offset,
-                                                            additional_offset,
-                                                            transducer_surface_model_filepath,
-                                                            focal_distance_list,
-                                                            flhm_list,
-                                                            placement_scene_template_filepath,
-                                                            ID=IDTarget) #if IDTarget is empty, the vertex number will be used
-                    else:
-                        print("No action taken.")
-
-                    # Reset the flag
-                    process_line = False
-
-    # Start the output reading in a separate thread
-    output_thread = threading.Thread(target=read_output)
-    output_thread.start()
-
-    # Wait for the process and threads to finish
-    process.wait()
-    output_thread.join()
-    listener.stop()
-    
-else:
-    # If only trajectory is to be done, call the appropriate function
-    PlanTUS.prepare_acoustic_simulation(args.do_only_trajectory,
-                                        output_path,
-                                        target_roi_filepath,
-                                        t1_filepath,
-                                        max_distance,
-                                        min_distance,
-                                        transducer_diameter,
-                                        max_angle,
-                                        plane_offset,
-                                        additional_offset,
-                                        transducer_surface_model_filepath,
-                                        focal_distance_list,
-                                        flhm_list,
-                                        placement_scene_template_filepath,
-                                        ID=IDTarget,
-                                        skip_wb_view=args.skip_wb_view)
