--- gitignore
@@ -1,2 +1,0 @@
-.DS_Store
-*.pyc
