+++ pyproject.toml
@@ -1,0 +1,35 @@
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[project]
+name = "PlanTUS"
+version = "0.1.1"
+authors = [
+  { name="Maximilian Lueckel", email="mlueckel@uni-mainz.de" },
+]
+description = "Tool for heuristic planning of TUS transducer placements"
+readme = "README.md"
+requires-python = ">=3.9"
+dependencies = [
+  "nibabel",
+  "nilearn", 
+  "numpy", 
+  "pynput", 
+  "PyQt5", 
+  "vtk",
+  "PyYAML"
+]
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "Operating System :: OS Independent",
+]
+license = "MIT"
+license-files = ["LICEN[CS]E*"]
+
+[project.scripts]
+plantus = "PlanTUS.PlanTUS_cli:main"
+
+[project.urls]
+Homepage = "https://github.com/mlueckel/PlanTUS"
+Issues = "https://github.com/mlueckel/PlanTUS/issues"
