+++ src/PlanTUS/__init__.py
@@ -1,0 +1,3 @@
+from PlanTUS.PlanTUS import *
+
+__all__ = ["PlanTUS"]
