+++ src/PlanTUS/resources/scene_templates/TUSTransducerPlacementPlanning_TEMPLATE.scene
