+++ src/PlanTUS/resources/scene_templates/TUSTransducerPlacement_TEMPLATE.scene
